Restructure parser code packages
