+++ parser/go.mod
@@ -4,5 +4,9 @@ go 1.15
 
 require (
 	github.com/Azure/azure-storage-blob-go v0.14.0
+	github.com/google/uuid v1.3.0 // indirect
 	github.com/streadway/amqp v1.0.0
+	golang.org/x/net v0.0.0-20211105192438-b53810dc28af // indirect
+	golang.org/x/sys v0.0.0-20211106132015-ebca88c72f68 // indirect
+	golang.org/x/text v0.3.7 // indirect
 )
