+++ parser/internal/formats/internalDiagnosticsFormats.go
