+++ parser/internal/service/commonparserlogic.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kozgot/go-log-processing/parser/pkg/formats"
+	"github.com/kozgot/go-log-processing/parser/internal/formats"
 	"github.com/kozgot/go-log-processing/parser/pkg/models"
 )
