+++ parser/internal/service/dcmessageparser.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kozgot/go-log-processing/parser/pkg/formats"
+	"github.com/kozgot/go-log-processing/parser/internal/formats"
 	"github.com/kozgot/go-log-processing/parser/pkg/models"
 )
