+++ parser/internal/service/inforparser.go
@@ -4,7 +4,7 @@ import (
 	"log"
 	"strings"
 
-	"github.com/kozgot/go-log-processing/parser/pkg/formats"
+	"github.com/kozgot/go-log-processing/parser/internal/formats"
 	"github.com/kozgot/go-log-processing/parser/pkg/models"
 )
