+++ parser/internal/service/smcConfigUpdateParser.go
@@ -3,7 +3,7 @@ package service
 import (
 	"strings"
 
-	"github.com/kozgot/go-log-processing/parser/pkg/formats"
+	"github.com/kozgot/go-log-processing/parser/internal/formats"
 	"github.com/kozgot/go-log-processing/parser/pkg/models"
 )
